+++ Code/LYRUIParticipantSectionHeaderView.h
@@ -1,5 +1,5 @@
 //
-//  LYRUIPaticipantSectionHeaderView.h
+//  LYRUIParticipantSectionHeaderView.h
 //  Pods
 //
 //  Created by Kevin Coleman on 8/31/14.
@@ -9,10 +9,10 @@
 #import <UIKit/UIKit.h>
 
 /**
- @abstract The `LYRUIPaticipantSectionHeaderView` displays a letter representing a section
+ @abstract The `LYRUIParticipantSectionHeaderView` displays a letter representing a section
  in the participant picker.
  */
-@interface LYRUIPaticipantSectionHeaderView : UIView
+@interface LYRUIParticipantSectionHeaderView : UIView
 
 /**
  @abstract Tells the receiver which letter to display for a given section.